changes to handle bugs

+++ app/src/main/java/com/example/aj/scavengersworld/DatabaseModels/UserProfile.java
@@ -6,14 +6,14 @@ import java.util.List;
  * Created by aj on 11/4/16.
  */
 public class UserProfile {
-    private int pointsEarned;
+    private double pointsEarned;
     private List<Integer> badgesEarned;
 
-    public int getPointsEarned() {
+    public double getPointsEarned() {
         return pointsEarned;
     }
 
-    public void setPointsEarned(int pointsEarned) {
+    public void setPointsEarned(double pointsEarned) {
         this.pointsEarned = pointsEarned;
     }
